Math functions return int in Python 3

+++ src/PIL/GifImagePlugin.py
@@ -25,6 +25,7 @@
 #
 
 import itertools
+import math
 import os
 import subprocess
 
@@ -701,14 +702,12 @@ def _get_optimize(im, info):
 
 def _get_color_table_size(palette_bytes):
     # calculate the palette size for the header
-    import math
-
     if not palette_bytes:
         return 0
     elif len(palette_bytes) < 9:
         return 1
     else:
-        return int(math.ceil(math.log(len(palette_bytes) // 3, 2))) - 1
+        return math.ceil(math.log(len(palette_bytes) // 3, 2)) - 1
 
 
 def _get_header_palette(palette_bytes):
